Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

So it takes a governor known to behave badly (including causing hung task panics in recovery and LPM) on our device, adds features to it which can be achieved in userspace (screen state scaling), and adds a SuperCoolMarketingName.

I don't think we want crap like this.  Every time I've seen someone include random-wacky-alternate-governor they've managed to achieve the Holy Fail of a device with poor performance AND poor battery life, or the true nature of their governor was masked by all the marketing bullshit (Lionheart).

Potentially look into seeing if the ondemand governor can have some of the tuning changes from conservative (such as reducing the minimum and default polling intervals) instead.  Maybe I'll mess with it this weekend.

Patch-set: 1