Update patch set 2
Patch Set 2:

Nope, I wasn't able to find a globally available accessory with this function. I'm pretty sure it is used in some car docks (see the spec), and, well, when I needed one I just made it.

It's easy: get a regular OTG cable and disassemble its phone micro-A end. Then, attach GND and VBUS to any 5V power source, and a 36k resistor between GND and ID. Voila, you made an ACA.

(For ACA to strictly conform to spec it needs to have a microcontroller inside and two FETs, but while you don't attach any OTG devices to it, as opposed to regular peripherals, this is actually a fully behaviorally conformant device).

If you don't want to go through it, I can just provide you a shell on my phone.

Why am I trying to get this into CM mainline? I have an (unreleased, because that'd be useless now) guide on making such a cable. Besides which, this change doesn't make the phone more active (charge pump is obviously not used in ACA mode), so I doubt this can ever have any negative consequence if working as expected.

Patch-set: 2