Update patch set 1
Patch Set 1:

"...known to.behave badly (including causing hung task panics in recovery and LPM)..."
Andrew, I had no clue about this. Can you give me a reference?

"...managed to achieve the Holy Fail of a device with poor performance AND poor battery life..."
I can't agree on that one. These things are really hard to measure since it really depends on how you are using your phone, but I have actually tested this governor for some days now and it suites my way of using my phone better than ondemand. Strange that a "random-wacky-alternate-governor" suites my way of using the phone better than the ondemand governor.

Maybe we should not use the smartassv2 governor as default (ondemand is still the default one, even if this change-set is used), but why not give the user the choice to try it - if there's no documented issues that can break your phone of course.

Patch-set: 1