Update patch set 3
Patch Set 3:

The problems with interactive were that it caused all sorts of stability issues when NOT selected or default.

I just saw a spontaneous reboot with last_kmsg symptoms similar to the stability problems experienced back during the last time interactive governor was compiled in.

If there are any more spontaneous resets today, I will be reverting this when I get home.

Patch-set: 3