Update patch set 1
Patch Set 1:

Andrew: I have now used conservative since you reported that smartassv2 might be a reason for spontaneous reboot when not used. I have still not had any troubles, but that does not prove that you are wrong and I'm still concerned about your findings. Yet I think smartassv2 suites me (and others) very well and until more testing with it loaded but not used, maybe a compromise is to build it as a module?

Patch-set: 1