Update patch set 3
Patch Set 3:

As long as the module is never loaded automatically that might be OK.

So far two spontaneous reboots.  In both cases, the last_kmsg is spammed with tons of kworker spam I've never seen before this patch was put in.  It's not as severe as interactive's problems were with kinteractived triggering a hung task panic (most likely due to this using workqueues instead of a separate task) when not default/selected, but it appears to be very similar.

There's no justification for this code to be in here anyway.  It achieves nothing that a combination of small tuning patches to existing governors (ondemand or conservative - and conservative has already been patched, ondemand probably needs a revisit) can't achieve when combined with screen-state profiles in userspace.

Patch-set: 3