Update patch set 1
Patch Set 1:

Seems to be a common theme that the GNex drivers are difficult to get working correctly on Hummingbird/Exynos devices.

Someone tried to grab pershoot's port of the tuna driver for use on Infuse - it sorta worked but not well.  Meanwhile once the prereqs in net/wireless were in place on Infuse (It's a 2.6.35 kernel so some backporting of the net/wireless framework was needed), drivers from Exynos devices could be dropped right in.

Patch-set: 1