rename ath6kl_sdio module to ath6kl
The exynos4tab device tree was originally configured with compat-wireless
which seperated out into ath6kl and ath6kl_sdio (like upstream linux). We
have to combine these two modules into one in order to really turn off the
wifi chip when disabled. Since ath6kl_sdio is just a feature for ath6kl, it
doesn't make sense to include it in the name for the whole module.

Change-Id: I7a6fe27e16b2213fd02bb71b3c739f375db17757