Update patch set 1
Patch Set 1:

So the question is - What are the extra six bytes on the last two values of the final filter entry on Tuna doing?

That filter is identical to the one we're reverting, except the last filter is:
104 0 0 0 0xFFFFFFFFFFFF 0x01005E0000FB
instead of
104 0 0 0 0xFFFFFF 0x01005E

I9100 ICS was a little different - slightly different filter settings for the first line than Gingerbread (but identical to Tuna settings and I think N7000 too), completely lacking the last line.  I've seen a number of complaints about battery life on some wifi networks with I9100 filter settings too.

Patch-set: 1