Update patch set 1
Patch Set 1:

I think the wacom driver should expose the tilt_offset values to user space. This is useful on the Galaxy Note, as a lot of people complain, that the S-Pen is off center. Changing tilt offsets allows to correct this.

I also created a patch for the GalaxyNoteSettings app (submitted to CyanogenMod/android_device_samsung_n7000) to allow normal users to do this configuration.

Patch-set: 1