tabs: disable Sec_Fifo DAI
The Sec_Fifo DAI doesn't get loaded while using the closed
source blobs. The opensource tinyalsa tries to start it
when certain routes are first initiated (such as the builtin
mic). Unfortunately loading this platform is currently broken
so this crashes audio, so we are disabling it (since it is
not necessary) for now.

Change-Id: I407005679a7613de292461dc8ac1531cdffbec00