Update patch set 2
Patch Set 2: I would prefer that you didn't submit this

Not necessary.

The commit claims that this is required to compile successfuly with GCC 4.6 - however, we removed TARGET_CUSTOM_TOOLCHAIN earlier this week and it's compiled fine.

So obviously the commit's claim that these are required is incorrect.

Cleanups to get rid of warnings often have unexpected consequences (see epic shittiness of Franco's N7000 kernels as a great example).

Patch-set: 2