Update patch set 2
Patch Set 2:

This commit is used for msm family of devices to successfully compile using GCC 4.6, but I removed the parts that weren't necessary as they were modifying parts for a msm processor. I've been using this since the merge of the JB update to the kernel for the d710 (epic4gtouch) with no negative consequences. I'll abandon if need be then.

Patch-set: 2